Fixed issue with unzipping nested zips
File was not closed before trying to use NSFileManager there it could not be opened to be extracted. Refactored the logic. Also skipped the steps of setting file properties if we plan to unzip to save unnecessary IO since the zips are deleted after extraction anyway.

Possibly in the failure section of unzippig the nested zip yuu would want to call the code to set the properties so at least the corrupt zip has proper attributes.