Set attributes on nested zip if extraction failure
As commented on prevoius commit. We normally skip setting attributes on nested zip sicne we inted to delete right after extraction. If there's an extraction error though, we go ahead and set the attributes. The intention being that the user may want to inspect the outcome for why there was an extraction error and we should restore attribute states.